% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{agg_covars}
\alias{agg_covars}
\title{get weighted covar cell totals for the given dataset,
covars, and weight var}
\usage{
agg_covars(dat, covars, weight)
}
\description{
get weighted covar cell totals for the given dataset,
covars, and weight var
}
