% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomtest_for_bin.R
\name{binom_for_histogram}
\alias{binom_for_histogram}
\title{binom_for_histogram}
\usage{
binom_for_histogram(n_x)
}
\arguments{
\item{n_x}{numeric vector of two integers. The first one is the number of cases in the bin; the second the number of instances in the bin}
}
\value{
p-value from stats::binom.test method
}
\description{
p_values from stats::binom.test for each bin, if bin is empty, a p-value of 2 is returned
}
