% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_BBQ.R
\name{build_BBQ}
\alias{build_BBQ}
\title{build_BBQ}
\usage{
build_BBQ(actual, predicted)
}
\arguments{
\item{actual}{vector of observed class labels (0/1)}

\item{predicted}{vector of uncalibrated predictions}
}
\value{
returns the BBQ model which includes models for all evaluated binning schemes; the prunedmodel contains only a selection of BBQ models with the best Bayesian score
}
\description{
This method builds a BBQ calibration model using the trainings set provided.
}
\details{
Based on the paper (and matlab code) : "Obtaining Well Calibrated Probabilities Using Bayesian Binning" by Naeini, Cooper and Hauskrecht: ; https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4410090/
}
