% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_GUESS.R
\name{build_GUESS}
\alias{build_GUESS}
\title{build_GUESS}
\usage{
build_GUESS(actual, predicted)
}
\arguments{
\item{actual}{vector of observed class labels (0/1)}

\item{predicted}{vector of uncalibrated predictions}
}
\value{
returns the trained GUESS model that can be used to calibrate a test set using the \code{\link{predict_GUESS}} method
}
\description{
This method builds a GUESS calibration model using the trainings set provided.
}
\seealso{
\code{\link[fitdistrplus]{denscomp}}
}
