% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_hist_binning.R
\name{build_hist_binning}
\alias{build_hist_binning}
\title{build_hist_binning}
\usage{
build_hist_binning(actual, predicted, bins = NULL)
}
\arguments{
\item{actual}{vector of observed class labels (0/1)}

\item{predicted}{vector of uncalibrated predictions}

\item{bins}{number of bins that should be used to build the binning model, Default: decide_on_break estimates optimal number of bins}
}
\value{
returns the trained histogram model that can be used to calibrate a test set using the \code{\link{predict_hist_binning}} method
}
\description{
calculate estimated probability per bin, input predicted and real score as numeric vector; builds a histogram binning model which can be used to calibrate uncalibrated predictions using the predict_histogramm_binning method
}
\details{
if trainings set is smaller then threshold (15 bins*5 elements=75), number of bins is decreased
}
