% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_me.R
\name{calibrate_me}
\alias{calibrate_me}
\title{calibrate_me}
\usage{
calibrate_me(actual, predicted, model_idx)
}
\arguments{
\item{actual}{vector of observed class labels (0/1)}

\item{predicted}{vector of uncalibrated predictions}

\item{model_idx}{a single number from 1 to 5, indicating which calibration model should be implemented, 1=hist_scaled, 2=hist_transformed, 3=BBQ_scaled, 4=BBQ_transformed, 5=GUESS}
}
\value{
depending on the value of \code{model_idx}, the respective calibration model is build on the input from \code{actual} and \code{predicted}
}
\description{
trains calibration models on the training set of \code{predicted}/\code{actual} value pairs.\code{model_idx} specifies which models should be trained.
}
