% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibratr-package.R
\docType{package}
\name{calibratr-package}
\alias{calibratr}
\alias{calibratr-package}
\title{calibratr: Calibration Metrics for Yardstick}
\description{
This package provides calibration metrics for use with the {yardstick} package, which is used to calculate performance metrics for models developed within the {Tidymodels} framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mglev1n/calibratr}
  \item Report bugs at \url{https://github.com/mglev1n/calibratr/issues}
}

}
\author{
\strong{Maintainer}: Michael Levin \email{mglev1n@gmail.com} (\href{https://orcid.org/0000-0002-9937-9932}{ORCID})

}
\keyword{internal}
