% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibratr}
\alias{calibratr}
\title{Help calibrate survey weights to population totals}
\usage{
calibratr(fml, pdat, pweight, sdesign, calfun = "linear")
}
\arguments{
\item{fml}{formula describing the calibration design (NB: this relies on
varnames and levels being the same in pdat and sdat)}

\item{pdat}{population-level data}

\item{pweight}{weights for adding up rows of population-level data}

\item{sdesign}{survey design object for survey dataset (see survey package)}

\item{calfun}{optional arg for survey package's calibrate fn}
}
\value{
a vector of calibrated weights
}
\description{
given the popn dataset, the survey dataset, and a formula describing
the calibration equation, produce calibration weights
}
\details{
Note that if you get warnings when running this, that may be a sign that the
survey data have missingness on some of the calibration variables. Missingness
will also cause total calibrated weights to differ from total un-calibrated weights.
}
