% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{compare_covar_margins}
\alias{compare_covar_margins}
\title{compare several different weights}
\usage{
compare_covar_margins(datalist, weights, names, covars, title)
}
\arguments{
\item{datalist}{a list of dataframes}

\item{weights}{a vector w/ names of weights,}

\item{names}{a vector w/ description of each data frame}

\item{covars}{to covariates to aggregate by}

\item{title}{}
}
\value{
A dataframe with different weight comparisons
}
\description{
Return a dataframe that has comparisons for different weights;
useful for checking different calibration specifications
}
