% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models_visual.R
\name{compare_models_visual}
\alias{compare_models_visual}
\title{compare_models_visual}
\usage{
compare_models_visual(models, seq = NULL)
}
\arguments{
\item{models}{PARAM_DESCRIPTION}

\item{seq}{sequence for which the calibrated predictions should be plotted, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{geom_line}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{ylim}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{labs}},\code{\link[ggplot2]{scale_color_brewer}}
 \code{\link[reshape2]{melt}}
}
