% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_values.R
\name{format_values}
\alias{format_values}
\title{format_values}
\usage{
format_values(cases, control, input, min = NULL, max = NULL, mean = NULL)
}
\arguments{
\item{cases}{instances from class 1}

\item{control}{instances from class 0}

\item{input}{single integer (0, 1 or 2). specify if the input should be formatted (=0), formatted and scaled (=1)
or formatted and transformed (=2)}

\item{min}{min value of the original data set, default=calculated on input}

\item{max}{max value of the original data set, default=calculated on input}

\item{mean}{mean value of the original data set, default=calculated on input}
}
\value{
list object with the following components:
\item{formated_values}{formatted input. If \code{input} is set to 1 (2), the input is additionally scaled (transformed) using the
method \code{\link{scale_me}} (\code{\link{transform_me}})}
\item{min}{minimum value among all instances}
\item{max}{maximum value among all instances}
\item{mean}{mean value among all instances}
}
\description{
returns formatted input.
If specified, the uncalibrated input is mapped to the [0;1] range using scaling (\code{\link{scale_me}}) or transforming (\code{\link{transform_me}})
}
