% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Brier_score.R
\name{get_Brier_score}
\alias{get_Brier_score}
\title{get_Brier_score}
\usage{
get_Brier_score(actual, predicted)
}
\arguments{
\item{actual}{vector of observed class labels (0/1)}

\item{predicted}{vector of uncalibrated predictions}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
