% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLES_in_comparison.R
\name{get_CLE_comparison}
\alias{get_CLE_comparison}
\title{get_CLE_comparison}
\usage{
get_CLE_comparison(list_models)
}
\arguments{
\item{list_models}{list object that contains all error values for all trained calibration models. For the specific format, see the calling function \code{\link{visualize_calibratR}}.}
}
\value{
ggplot2
}
\description{
visualises how class 1 and class 0 classification error (CLE) differs in each trained calibration model.
Comparing class-specific CLE helps to choose a calibration model for applications were classification error is cost-sensitive for one class.
See \code{\link{get_CLE_class}} for details on the implementation.
}
