% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getECE.R
\name{get_ECE_equal_width}
\alias{get_ECE_equal_width}
\title{get_ECE_equal_width}
\usage{
get_ECE_equal_width(actual, predicted, bins = 10)
}
\arguments{
\item{actual}{vector of observed class labels (0/1)}

\item{predicted}{vector of uncalibrated predictions}

\item{bins}{number of bins for the equal-width binning model}
}
\value{
equal-width ECE value
}
\description{
Expected Calibration Error (ECE); the model is divided into  10 equal-width bins (default) and the mean of the observed (0/1) vs. mean of predicted is calculated per bin, weighted by emperical frequency of elements in bin i
}
