% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMCE.R
\name{get_MCE_equal_width}
\alias{get_MCE_equal_width}
\title{get_MCE_equal_width}
\usage{
get_MCE_equal_width(actual, predicted, bins = 10)
}
\arguments{
\item{actual}{vector of observed class labels (0/1)}

\item{predicted}{vector of uncalibrated predictions}

\item{bins}{number of bins for the binning model}
}
\value{
equal-width MCE value
}
\description{
Maximum Calibration Error (MCE), returns maximum calibration error for equal-width binning model
}
