% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getECE.R
\name{getECE}
\alias{getECE}
\title{getECE}
\usage{
getECE(actual, predicted, n_bins = 10)
}
\arguments{
\item{actual}{vector of observed class labels (0/1)}

\item{predicted}{vector of uncalibrated predictions}

\item{n_bins}{number of bins of the underlying equal-frequency histogram, Default: 10}
}
\value{
equal-frequency ECE value
}
\description{
Expected Calibration Error (ECE); the model is divided into 10 equal-width bins (default) and the mean of the observed (0/1) vs. mean of predicted is calculated per bin, weighted by emperical frequency of elements in bin i
}
