% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRMSE.R
\name{getRMSE}
\alias{getRMSE}
\title{getRMSE}
\usage{
getRMSE(actual, predicted)
}
\arguments{
\item{actual}{vector of observed class labels (0/1)}

\item{predicted}{vector of uncalibrated predictions}
}
\value{
RMSE value
}
\description{
calculates the root of mean square error (RMSE) in the test set of calibrated predictions
}
