% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{plot_model}
\alias{plot_model}
\title{plot_model}
\usage{
plot_model(calibration_model, seq = NULL)
}
\arguments{
\item{calibration_model}{output from the \code{\link{calibrate}} method.}

\item{seq}{sequence of ML scores over which the mapping function should be evaluated, Default: 100 scores from the minimum to the maximum of the original ML scores}
}
\value{
ggplot object
}
\description{
this methods visualizes all implemented calibration models as a mapping function between original ML scores (x-axis) and
calibrated predictions (y-axis)
}
\seealso{
\code{\link[reshape2]{melt}}
 \code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{geom_line}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{ylim}},\code{\link[ggplot2]{scale_colour_manual}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{labs}},\code{\link[ggplot2]{geom_text}},\code{\link[ggplot2]{geom_vline}}
}
