% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_BBQ.R
\name{predict_BBQ}
\alias{predict_BBQ}
\title{predict_BBQ}
\usage{
predict_BBQ(bbq, new, option)
}
\arguments{
\item{bbq}{output from the \code{\link{build_BBQ}} method}

\item{new}{vector of uncalibrated probabilities}

\item{option}{either 1 or 0; averaging=1, selecting=0}
}
\value{
a list object containing the following components:
\item{predictions}{contains a vector of calibrated predictions}
\item{pred_idx}{which option was used (averaging or selecting)}
\item{significance_test_set}{the percentage of \code{new} instances that was evaluated using significant prediction estimates}
\item{pred_per_bin}{number of instances \code{new} in each bin of the selected model}
}
\description{
FUNCTION_DESCRIPTION
}
\details{
Based on the paper (and matlab code) : "Obtaining Well Calibrated Probabilities Using Bayesian Binning" by Naeini, Cooper and Hauskrecht: ; https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4410090/
}
