% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_hist_binning.R
\name{predict_hist_binning}
\alias{predict_hist_binning}
\title{predict_hist_binning}
\usage{
predict_hist_binning(histogram, new)
}
\arguments{
\item{histogram}{the output of \code{\link{build_hist_binning}}}

\item{new}{vector of uncalibrated probabilities}
}
\value{
a list object containing the following components
\item{predictions}{contains a vector of calibrated predictions}
\item{significance_test_set}{the percentage of \code{new} instances that was evaluated using significant prediction estimates}
\item{pred_per_bin}{a table containing the number of instances from \code{new} for each bin of the final binning scheme of \code{histogram}}
}
\description{
predict for a new element using histogram binning
}
