% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_model.R
\name{predict_model}
\alias{predict_model}
\title{predict_model}
\usage{
predict_model(new, calibration_model, min, max, mean, inputtype)
}
\arguments{
\item{new}{vector of uncalibrated predictions}

\item{calibration_model}{calibration model to be used for the calibration. Can be the output of \code{\link{build_BBQ}},\code{\link{build_hist_binning}} or \code{\link{build_GUESS}}.}

\item{min}{minimum value of the original data set}

\item{max}{maximum value of the original data set}

\item{mean}{mean value of the original data set}

\item{inputtype}{specify if the model was build on original (=0), scaled(=1) or transformed (=2) data}
}
\value{
vector of calibrated predictions
}
\description{
calibrates the uncalibrated predictions \code{new} using \code{calibration_model}.
}
