% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_multiple_runs.R
\name{rd_multiple_runs}
\alias{rd_multiple_runs}
\title{rd_multiple_runs}
\usage{
rd_multiple_runs(list_models)
}
\arguments{
\item{list_models}{list object that contains n-times the output from the \code{\link{reliability_diagramm}}. method.}
}
\value{
a list object that contains a reliability diagram that visualises all reliabilty diagrams that were constructed during n-times repeated m-fold cross-validation.
}
\description{
This functions plots all n reliability diagrams that were constructed during n-times repeated m-fold cross-validation (CV).
During calibration model evaluation, CV is repeated n times, so that eventually n reliability diagrams are obtained.
}
\seealso{
\code{\link[reshape2]{melt}}
 \code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{geom_line}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_abline}},\code{\link[ggplot2]{ylab}},\code{\link[ggplot2]{xlab}},\code{\link[ggplot2]{xlim}},\code{\link[ggplot2]{ylim}},\code{\link[ggplot2]{coord_fixed}},\code{\link[ggplot2]{geom_text}},\code{\link[ggplot2]{scale_color_discrete}},\code{\link[ggplot2]{ggtitle}}
}
