% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability_diagram.R
\name{reliability_diagramm}
\alias{reliability_diagramm}
\title{reliability_diagramm}
\usage{
reliability_diagramm(actual, predicted, bins = 10, plot_rd = TRUE)
}
\arguments{
\item{actual}{vector of observed class labels (0/1)}

\item{predicted}{vector of uncalibrated predictions}

\item{bins}{number of bins in the reliability diagram, Default: 10}

\item{plot_rd}{should the reliability diagram be plotted, Default: TRUE}
}
\value{
a list object containing the following elements
\item{calibration_error}{}
\item{discrimination_error}{}
\item{rd_breaks}{}
\item{histogram_plot}{}
\item{diagram_plot}{}
\item{mean_pred_per_bin}{}
\item{accuracy_per_bin}{}
\item{freq_per_bin}{}
\item{sign}{}
}
\description{
Reliability curves allow checking if the predicted probabilities of a
}
\seealso{
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{stat_bin}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{scale_fill_manual}},\code{\link[ggplot2]{theme}},\code{\link[ggplot2]{labs}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{xlim}},\code{\link[ggplot2]{ylim}},\code{\link[ggplot2]{geom_abline}},\code{\link[ggplot2]{geom_line}},\code{\link[ggplot2]{geom_text}},\code{\link[ggplot2]{geom_label}},\code{\link[ggplot2]{coord_fixed}}
}
