% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_values.R
\name{transform_me}
\alias{transform_me}
\title{transform_me}
\usage{
transform_me(x_unscaled, mean)
}
\arguments{
\item{x_unscaled}{vector of predictions}

\item{mean}{mean of \code{x}}
}
\value{
transformed values of \code{x_unscaled}
}
\description{
maps all instances in \code{x_unscaled} to the [0;1] range using the equation:
\cr  y=exp(x)/(1+exp(x))
}
\details{
values greater then exp(700)/ or smaller then exp(-700) are returned as "Inf". To avoid NaN values, these "Inf." values are turned into min(y) or max(y).
}
