% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_calibrated_test_set.R
\name{visualize_calibrated_test_set}
\alias{visualize_calibrated_test_set}
\title{visualize_calibrated_test_set}
\usage{
visualize_calibrated_test_set(actual, predicted_list, cutoffs)
}
\arguments{
\item{actual}{vector of observed class labels (0/1)}

\item{predicted_list}{predict_calibratR$predictions object (list of calibrated predictions from calibration models)}

\item{cutoffs}{vector of optimal cut-off thresholds for each calibration model}
}
\value{
ggplot2 element for visual comparison of the evaluated calibration models
}
\description{
plots a panel for all calibrated predictions from the respective calibration model. Allows visual comparison of the models output and their optimal cut off
}
\seealso{
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{scale_colour_manual}},\code{\link[ggplot2]{xlab}},\code{\link[ggplot2]{ylab}},\code{\link[ggplot2]{geom_hline}},\code{\link[ggplot2]{ylim}}
}
